Move terrain routes into terrainHandler.js
