+++ src/server/terrainHandler.js
@@ -2,9 +2,11 @@ import request from "request";
 import path from "path";
 import querystring from "querystring";
 import { terrainURL } from "./configuration";
+import * as authStrategy from "./authStrategy";
 import logger from "./logging";
+import express from "express";
 
-export const terrain = ({ method, url, pathname, headers }) => {
+const terrain = ({ method, url, pathname, headers }) => {
     const baseURL = url || terrainURL;
 
     return (req, res) => {
@@ -28,3 +30,60 @@ export const terrain = ({ method, url, pathname, headers }) => {
         req.pipe(request(requestOptions)).pipe(res);
     };
 };
+
+export default function terrainRouter() {
+    logger.info("creating the api router");
+    const api = express.Router();
+
+    logger.info("adding the /api/upload handler ");
+    api.post(
+        "/upload",
+        authStrategy.authnTokenMiddleware,
+        terrain({
+            method: "POST",
+            pathname: "/secured/fileio/upload",
+        })
+    );
+
+    logger.info("adding the /api/download handler");
+    api.get(
+        "/download",
+        authStrategy.authnTokenMiddleware,
+        terrain({
+            method: "GET",
+            pathname: "/secured/fileio/download",
+        })
+    );
+
+    logger.info("adding the /api/filesystem/paged-directory handler");
+    api.get(
+        "/filesystem/paged-directory",
+        authStrategy.authnTokenMiddleware,
+        terrain({
+            method: "GET",
+            pathname: "/secured/filesystem/paged-directory",
+        })
+    );
+
+    logger.info("adding the /api/filesystem/stat handler");
+    api.post(
+        "/filesystem/stat",
+        authStrategy.authnTokenMiddleware,
+        terrain({
+            method: "POST",
+            pathname: "/secured/filesystem/stat",
+            headers: {
+                "Content-Type": "application/json",
+            },
+        })
+    );
+
+    logger.info("adding the /api/filesystem/root handler");
+    api.get(
+        "/filesystem/root",
+        authStrategy.authnTokenMiddleware,
+        terrain({ method: "GET", pathname: "/secured/filesystem/root" })
+    );
+
+    return api;
+}
